pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cleanmap"
version = "0.1.0"
description = "CLI tool for cleaning images"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pillow>=10.0",
    "google-genai>=0.8",
    "python-dotenv>=1.0",
]

[project.scripts]
cleanmap = "cleanmap.cli:main"
